🐛 fix(governance): round 3 — fix TreasurySpend panic and multi-sig stuck state
- Fix TreasurySpend execution_data length check: <48 allowed 48–63 byte
  inputs to panic on execution_data[48..64]. Now requires <64.
- Fix emergency pause/resume multi-sig: record signatures only after
  verifying chain state (paused/not paused), preventing stuck multi-sig
  state when threshold is reached but the operation is invalid.
- Add regression tests for 50-byte TreasurySpend and pause/resume
  idempotency guards.

+++ crates/governance/src/config.rs
@@ -76,7 +76,8 @@ impl Default for GovernanceConfig {
 impl GovernanceConfig {
     /// Calculate validator quorum count (ceil of total_validators * bps / 10000)
     pub fn validator_quorum(&self, total_validators: u64) -> u64 {
-        (total_validators * self.validator_quorum_bps as u64).div_ceil(10_000)
+        ((total_validators as u128) * (self.validator_quorum_bps as u128))
+            .div_ceil(10_000) as u64
     }
 
     /// Calculate supply quorum (ceil of total_supply * bps / 10000)
@@ -91,11 +92,13 @@ impl GovernanceConfig {
 
     /// Calculate simple majority quorum count
     pub fn simple_majority(&self, total_validators: u64) -> u64 {
-        (total_validators * self.simple_majority_bps as u64).div_ceil(10_000)
+        ((total_validators as u128) * (self.simple_majority_bps as u128))
+            .div_ceil(10_000) as u64
     }
 
     /// Calculate emergency pause threshold
     pub fn emergency_pause_threshold(&self, total_validators: u64) -> u64 {
-        (total_validators * self.emergency_pause_bps as u64).div_ceil(10_000)
+        ((total_validators as u128) * (self.emergency_pause_bps as u128))
+            .div_ceil(10_000) as u64
     }
 }
